+++ forui/lib/src/widgets/autocomplete/autocomplete_content.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
@@ -146,7 +147,7 @@ class Content extends StatelessWidget {
 }
 
 /// An [FAutocomplete]'s content style.
-class FAutocompleteContentStyle with Diagnosticable, _$FAutocompleteContentStyleFunctions {
+class FAutocompleteContentStyle extends FPopoverStyle with Diagnosticable, _$FAutocompleteContentStyleFunctions {
   /// The default text style when there are no results.
   @override
   final TextStyle emptyTextStyle;
@@ -168,14 +169,18 @@ class FAutocompleteContentStyle with Diagnosticable, _$FAutocompleteContentStyle
     required this.emptyTextStyle,
     required this.progressStyle,
     required this.sectionStyle,
+    required super.decoration,
     this.padding = const EdgeInsets.symmetric(vertical: 5),
+    super.barrierFilter,
+    super.backgroundFilter,
+    super.viewInsets,
   });
 
   /// Creates a [FAutocompleteContentStyle] that inherits its properties.
-  FAutocompleteContentStyle.inherit({required FColors colors, required FTypography typography, required FStyle style})
-    : this(
-        emptyTextStyle: typography.sm,
-        progressStyle: FCircularProgressStyle.inherit(colors: colors),
-        sectionStyle: FAutocompleteSectionStyle.inherit(colors: colors, style: style, typography: typography),
-      );
+  FAutocompleteContentStyle.inherit({required super.colors, required FTypography typography, required super.style})
+    : emptyTextStyle = typography.sm,
+      progressStyle = FCircularProgressStyle.inherit(colors: colors),
+      sectionStyle = FAutocompleteSectionStyle.inherit(colors: colors, style: style, typography: typography),
+      padding = const EdgeInsets.symmetric(vertical: 5),
+      super.inherit();
 }