+++ forui/lib/src/widgets/select/single/search_select.dart
@@ -43,7 +43,7 @@ class _SearchSelect<T> extends FSelect<T> {
     super.fieldAnchor,
     super.popoverConstraints,
     super.spacing,
-    super.shift,
+    super.overflow,
     super.offset,
     super.hideRegion,
     super.autoHide,