+++ forui/lib/src/widgets/time_field/time_field.dart
@@ -300,8 +300,8 @@ abstract class FTimeField extends StatefulWidget {
   /// The [spacing] property controls the spacing between the input field and the picker popover. Defaults to
   /// `FPortalSpacing(4)`.
   ///
-  /// The [shift] function controls how the picker repositions when space is constrained. Defaults to
-  /// [FPortalShift.flip].
+  /// The [overflow] function controls how the picker repositions when space is constrained. Defaults to
+  /// [FPortalOverflow.flip].
   ///
   /// The [offset] property controls the offset of the picker popover. Defaults to [Offset.zero].
   ///
@@ -331,7 +331,7 @@ abstract class FTimeField extends StatefulWidget {
     Alignment anchor,
     Alignment inputAnchor,
     FPortalSpacing spacing,
-    Offset Function(Size, FPortalChildBox, FPortalBox) shift,
+    FPortalOverflow overflow,
     Offset offset,
     FPopoverHideRegion hideRegion,
     VoidCallback? onTapHide,